requirements.txt
# Core Framework
fastapi==0.121.2
uvicorn[standard]==0.38.0
starlette==0.49.3

# Database
SQLAlchemy==2.0.44
psycopg[binary]==3.2.12
alembic==1.15.0

# API & External Services
openai==2.8.0
perplexityai==0.20.0
requests==2.32.5
httpx==0.28.1

# Utilities
python-dotenv==1.2.1
pydantic==2.12.4
email-validator==2.3.0
python-multipart==0.0.20

# Development & Linting
ruff==0.14.8